More documentation and deprecate birch
