+++ examples/simple/model/dwarf_diffuse.png
