+++ view/model.go
@@ -7,6 +7,8 @@ import (
 	"github.com/oakmound/oak/render"
 )
 
+// Model represents a 3D model in world space.
+// This model is a render.Renderable.
 type Model struct {
 	// a render.Sprite has a position and a buffer of image data which
 	// it uses to draw to the screen at that position.
@@ -24,22 +26,61 @@ type Model struct {
 	// applied to the camera's viewing position and angles to create
 	// the illusion of full 3D rotation.
 	quat mgl64.Quat
+
+	scale    mgl64.Mat4
+	position mgl64.Mat4
 	// transform represents the model's position and scale.
-	transform mgl64.Mat4
-	angle     float64
+	// transform mgl64.Mat4
+	angle float64
 
 	camera *Camera
 }
 
-// Rotate resets the rotation of the model to what is provided.
+// SetRotation resets the rotation of the model to what is provided.
 // If you wish to rotate based on the current rotation then please refer to
-// RotateExisting instead.
-func (m *Model) Rotate(angle float64, axis mgl64.Vec3) {
+// AddRotation instead.
+func (m *Model) SetRotation(angle float64, axis mgl64.Vec3) {
 	m.angle = angle
 	m.quat = mgl64.QuatRotate(angle, axis)
 }
 
-func (m *Model) RotateExisting(angle float64, axis mgl64.Vec3) {
+// AddRotation adds to the existing rotation axis.
+func (m *Model) AddRotation(angle float64, axis mgl64.Vec3) {
 	m.angle += angle
 	m.quat = mgl64.QuatRotate(m.angle, axis)
+}
+
+// GetTransform combines the scale and position to give the transform matrix of
+// this model.
+func (m *Model) GetTransform() mgl64.Mat4 {
+	return m.scale.Mul4(m.position)
+}
+
+// GetScale gets the model's scale on its x, y, and z axis.
+func (m *Model) GetScale() mgl64.Vec3 {
+	return m.scale.Diag().Vec3()
+}
+
+// SetScale sets the model's relative scale on the x, y, and z axis
+// If you wish to scale based on thge current rotation then please refer to
+// AddScale instead.
+func (m *Model) SetScale(x, y, z float64) {
+	m.scale = mgl64.Scale3D(x, y, z)
+}
+
+// AddScale scales the object relative to its current scale.
+func (m *Model) AddScale(x, y, z float64) {
+	m.scale = m.scale.Add(mgl64.Scale3D(x, y, z))
+}
+
+// SetPosition sets the position of the object from 0,0,0.
+// If you wish to set the scale based on its current position then please refer
+// to AddPosition instead.
+func (m *Model) SetPosition(x, y, z float64) {
+	m.position = mgl64.Translate3D(x, y, z)
+}
+
+// AddPosition sets the model's position relative to its current position.
+func (m *Model) AddPosition(x, y, z float64) {
+	m.position = m.position.Add(mgl64.Translate3D(x, y, z))
 }
